added processing of incorrect commands in CryptoBot.java and ServiceCommands enum

+++ src/main/java/com/skillbox/cryptobot/bot/command/GetSubscriptionCommand.java
@@ -32,7 +32,7 @@ public class GetSubscriptionCommand implements IBotCommand {
         SendMessage answer = new SendMessage();
         answer.setChatId(message.getChatId());
 
-        Long price = service.getSubscription(message.getChatId());
+        Double price = service.getSubscription(message.getChatId());
         if (price == null) {
             answer.setText("Активные подписки отсутствуют");
         } else {
